Use PSR-4 instead of PSR-0
Also PSR-4:d the tests
