+++ composer.json
@@ -1,19 +1,31 @@
 {
-  "name": "airbrake/phpbrake",
-  "description": "Airbrake exception and error notifier for PHP",
-  "keywords": ["airbrake", "notifier", "exception", "error", "logging"],
-  "homepage": "https://airbrake.io",
-  "license": "MIT",
-  "require-dev": {
-    "phpunit/phpunit": "^4.7",
-    "monolog/monolog": "^1.15"
-  },
-  "autoload": {
-    "psr-0": {
-      "Airbrake\\": "src/"
-    }
-  },
-  "config": {
-    "bin-dir": "bin"
-  }
+    "name": "airbrake/phpbrake",
+    "description": "Airbrake exception and error notifier for PHP",
+    "keywords": ["airbrake", "notifier", "exception", "error", "logging"],
+    "homepage": "https://airbrake.io",
+    "license": "MIT",
+    "require": {
+        "php": ">=5.3.3"    
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^4.8.0",
+        "monolog/monolog": "^1.16.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Airbrake\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Airbrake\\Tests\\": "tests/"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0-dev"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
