+++ src/ErrorHandler.php
