+++ src/Errors/Base.php
